src/requirements.txt
amqp
anyio
appnope
APScheduler
argon2-cffi
argon2-cffi-bindings
asgiref
asttokens
attrs
autopep8
backcall
Beaker
beautifulsoup4
billiard
bleach
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
comm
contourpy
cron-descriptor
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
distlib
dj-rest-auth
Django
django-admin-volt
django-allauth
django-bootstrap-modal-forms
django-allauth[socialaccount]
django-celery-beat
django-celery-results
django-cors-headers
django-countries
django-crispy-forms
django-debug-toolbar
django-dotenv
django-environ
django-filter
django-ipware
django-quill-editor
django-rest-auth
django-rest-passwordreset
django-storages
django-timezone-field
djangorestframework
ecdsa
entrypoints
et-xmlfile
executing
fastjsonschema
filelock
fonttools
future
hashids
idna
importlib-metadata
infobip-api-python-sdk
instamojo-wrapper
ipykernel
ipython
ipython-genutils
jedi
Jinja2
jsonschema
jupyter-client
jupyter-server
jupyter_core
jupyterlab-pygments
kiwisolver
kombu
Mako
MarkupSafe
marshmallow
marshmallow-enum
marshmallow-jsonapi
matplotlib
matplotlib-inline
mistune
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
oauthlib
oic
openpyxl
packaging
pandas
pandocfilters
parso
pep8
pexpect
pickleshare
Pillow
platformdirs
prometheus-client
prompt-toolkit
psutil
psycopg2-binary
ptyprocess
pure-eval
pyasn1
pycodestyle
pycparser
pycryptodome
pycryptodomex
pydantic
pyfcm
Pygments
pyjwkest
PyJWT
pyparsing
pyrsistent
python-crontab
python-dateutil
python-decouple
python-dotenv
python-jose
python-json-logger
python3-openid
pytz
pyzmq
razorpay
regex
requests
requests-oauthlib
rsa
Send2Trash
six
sniffio
soupsieve
sqlparse
stack-data
stripe
terminado
tinycss2
tornado
traitlets
typing-inspect
typing_extensions
tzdata
tzlocal
uri-template
urllib3
vine
virtualenv
wcwidth
webcolors
webencodings
websocket-client
whitenoise
xlrd
zipp
drf_yasg
ua-parser
django-user-agents
redis